+++ Estructuras de Datos Avanzadas/Las Pistas de Azul/.gitignore
@@ -1,0 +1,2 @@
+**/*.out
+!tests/invalid-cases/*.out
