+++ Estructuras de Datos Avanzadas/Las Pistas de Azul/cases/caso1.1.in
@@ -1,4 +1,5 @@
-4 2 2
+4
+2 2
 2 1 2 0 2 2 0 2
 3 2 1 3 0 2 3 0
 3 1 3 0 -1 -1 0 1
