+++ Estructuras de Datos Avanzadas/Las Pistas de Azul/cases/caso1.1.out
