+++ Estructuras de Datos Avanzadas/Las Pistas de Azul/cases/caso1.2.in
@@ -1,4 +1,5 @@
-10 7 2
+10
+7 2
 7 9 0 8 3 6 1 5 7 0 4 1 4 3 4 4 3 5 6 5
 9 7 9 8 7 7 3 2 4 6 0 4 8 0 8 8 2 4 8 2
 0 7 5 4 1 9 3 3 6 8 4 7 1 4 5 6 4 5 2 7
