+++ Estructuras de Datos Avanzadas/Las Pistas de Azul/cases/caso1.2.out
