+++ Estructuras de Datos Avanzadas/Las Pistas de Azul/cases/caso2.1.in
@@ -1,4 +1,5 @@
-14 2 6
+14
+2 6
 7 5 3 10 5 5 9 12 10 13 5 3 11 11 7 11 1 0 10 1 11 12 1 5 7 7 10 6
 11 6 12 5 1 10 12 1 12 11 0 2 12 2 2 12 8 1 5 2 6 5 6 0 5 8 2 5
 8 8 7 1 9 13 11 1 5 10 4 4 -1 -1 0 3 1 9 6 11 1 13 13 1 12 0 3 4
