+++ Estructuras de Datos Avanzadas/Las Pistas de Azul/cases/caso2.1.out
