+++ Estructuras de Datos Avanzadas/Las Pistas de Azul/cases/caso2.2.in
@@ -1,4 +1,5 @@
-14 11 0
+14
+11 0
 0 5 13 10 2 0 4 3 8 8 9 8 0 8 11 11 7 6 10 9 11 4 13 9 12 0 0 6
 6 1 7 1 1 5 5 8 3 12 10 8 8 0 6 13 4 6 3 10 9 2 7 2 3 13 6 8
 2 11 2 10 1 4 11 9 7 11 8 5 11 8 9 9 5 10 10 5 6 0 8 11 13 2 12 8
