+++ Estructuras de Datos Avanzadas/Las Pistas de Azul/cases/caso3.1.in
@@ -1,4 +1,5 @@
-46 24 17
+46
+24 17
 6 3 11 0 44 7 3 10 25 4 6 6 12 29 33 13 26 12 45 29 38 10 18 15 11 40 28 4 5 10 34 18 30 16 37 5 28 12 22 40 39 34 40 35 34 4 31 16 23 23 27 31 5 19 20 36 38 22 34 11 35 28 16 20 39 18 33 29 10 41 3 6 25 18 11 42 38 17 36 37 0 36 9 9 24 15 22 9 14 19 28 13
 44 33 27 22 45 28 15 19 2 35 25 20 2 4 28 18 26 9 39 1 20 16 33 3 28 20 2 21 3 19 3 39 36 5 7 15 1 7 16 33 39 30 33 37 14 27 18 1 15 26 28 14 34 28 0 10 34 3 35 13 10 27 35 22 9 10 11 36 2 27 31 4 35 38 44 39 28 24 4 39 13 44 27 30 45 9 16 34 41 42 44 43
 25 16 18 41 37 45 33 12 41 0 14 12 8 18 37 20 23 45 37 3 34 36 13 2 2 14 10 20 13 31 25 21 3 24 5 16 43 41 17 12 38 34 33 44 16 24 30 44 12 40 2 22 41 3 14 15 5 3 29 38 1 43 40 27 9 27 13 32 29 40 19 24 14 29 3 36 10 31 35 11 31 32 11 19 32 26 21 19 45 14 0 9
