+++ Estructuras de Datos Avanzadas/Las Pistas de Azul/cases/caso3.1.out
