+++ Estructuras de Datos Avanzadas/Las Pistas de Azul/cases/caso3.2.in
@@ -1,4 +1,5 @@
-50 28 8
+50
+28 8
 25 27 34 14 18 36 20 0 41 30 31 30 43 36 36 9 30 30 44 35 29 48 49 15 32 12 3 13 7 14 23 22 4 10 36 47 0 20 40 30 10 45 17 46 32 20 12 45 35 11 41 42 35 49 30 37 33 47 25 7 18 3 6 47 9 23 25 18 44 12 4 42 15 42 32 9 23 1 44 5 47 22 27 8 26 17 4 22 24 29 22 39 24 47 31 20 19 32 48 23
 47 30 40 1 31 47 15 43 24 12 8 26 14 8 46 22 19 37 28 7 29 19 29 9 46 8 5 13 9 13 39 24 36 38 14 9 40 25 36 2 37 41 18 46 40 34 28 23 2 12 3 11 41 9 19 29 3 0 16 31 5 32 20 42 11 14 31 16 40 9 28 2 17 28 38 1 9 32 12 19 35 43 42 41 33 27 38 20 40 27 35 42 11 2 3 8 28 1 5 3
 3 5 0 30 12 6 38 23 39 43 13 49 19 18 26 30 23 36 42 11 33 49 12 5 0 5 3 28 21 42 8 9 21 4 23 38 8 16 6 30 28 4 33 36 11 16 33 1 8 33 12 31 0 19 44 6 17 0 48 44 47 36 19 4 44 17 32 38 20 15 29 11 5 26 17 35 31 39 41 1 48 17 0 33 14 10 39 6 15 13 12 23 21 47 12 8 39 13 47 31
