+++ Estructuras de Datos Avanzadas/Las Pistas de Azul/cases/caso4.1.in
@@ -1,4 +1,5 @@
-50 10 36
+50
+10 36
 39 31 13 20 4 26 25 18 6 13 18 42 37 3 29 34 5 1 28 17 28 14 29 13 49 15 29 47 45 39 7 28 30 6 38 37 23 12 1 21 0 36 27 43 15 8 5 7 32 16 40 28 46 34 6 40 29 14 20 46 18 33 23 46 26 46 32 2 35 25 43 10 5 4 23 2 5 26 11 22 25 8 46 1 27 9 9 16 9 30 7 20 37 18 28 26 36 29 10 22
 17 2 12 42 15 24 45 43 16 4 28 25 20 24 20 8 37 38 48 3 12 42 36 5 12 4 49 24 23 37 33 9 2 21 16 15 11 27 41 18 27 47 48 28 8 16 44 11 41 29 29 47 45 17 11 24 24 27 11 41 9 22 40 36 27 1 24 37 28 32 12 15 38 18 46 24 35 46 31 15 29 45 15 6 5 29 21 11 7 24 0 45 28 21 47 9 43 1 17 39
 6 18 29 36 49 20 37 23 13 43 37 25 23 6 21 8 4 9 11 35 13 30 31 37 38 16 18 33 47 39 33 22 33 21 12 15 17 28 47 33 41 4 28 49 16 43 16 8 28 12 22 46 33 16 9 2 30 21 44 30 14 20 29 20 39 48 25 37 48 30 35 10 41 25 18 24 4 42 34 0 47 3 8 13 43 11 40 34 3 24 40 5 9 10 2 12 19 1 9 48
