+++ Estructuras de Datos Avanzadas/Las Pistas de Azul/cases/caso4.1.out
