+++ Estructuras de Datos Avanzadas/Las Pistas de Azul/cases/caso4.2.in
@@ -1,4 +1,5 @@
-50 42 2
+50
+42 2
 47 22 1 27 16 32 23 9 12 32 7 16 7 22 44 43 22 16 44 13 33 45 4 1 25 31 4 32 22 35 22 14 26 25 17 30 3 13 41 12 16 5 44 0 23 46 43 44 30 44 35 45 24 4 30 14 40 35 37 4 30 47 3 48 37 27 43 43 1 34 26 42 34 1 40 40 29 32 32 12 14 37 18 27 45 39 31 48 12 36 12 37 34 44 26 43 10 26 46 8
 38 49 28 11 20 8 34 48 3 32 24 26 38 18 8 3 23 33 28 13 44 37 3 15 28 37 13 18 14 8 43 3 3 2 9 21 31 38 8 11 18 19 19 6 11 41 17 41 39 30 8 0 35 27 27 39 28 35 34 4 46 29 30 13 39 5 1 40 34 13 16 47 36 32 44 19 49 5 29 37 14 39 27 20 5 18 34 16 43 36 11 12 6 27 21 4 48 13 31 24
 20 7 8 32 13 36 29 41 21 36 20 49 31 45 22 42 31 3 25 39 5 38 46 27 0 2 45 17 5 25 43 31 22 32 16 42 38 25 34 14 42 4 38 31 41 26 43 32 32 17 48 34 18 32 25 34 15 14 27 28 24 34 47 23 44 28 5 24 34 3 36 17 29 11 42 11 25 24 10 12 33 8 18 1 16 21 32 23 23 33 13 13 31 12 19 34 14 1 22 45
