+++ Estructuras de Datos Avanzadas/Las Pistas de Azul/cases/caso4.2.out
