+++ Estructuras de Datos Avanzadas/Las Pistas de Azul/cases/caso5.1.in
@@ -1,4 +1,5 @@
-23 22 15
+23
+22 15
 15 20 20 10 13 20 8 21 3 7 11 13 13 0 21 21 10 22 10 6 13 22 8 1 4 8 18 16 13 8 6 14 18 11 15 6 19 5 5 8 8 2 11 18 21 18
 22 10 9 18 3 4 22 15 0 22 10 0 12 19 6 0 20 2 21 16 22 14 19 7 12 22 7 12 11 9 22 12 16 16 5 18 19 4 0 9 10 8 8 0 15 16
 17 7 14 7 14 15 3 4 20 19 0 1 20 19 11 5 15 8 3 20 21 7 9 18 12 7 0 14 12 14 1 9 6 1 5 17 9 10 1 20 19 1 22 17 18 15
