+++ Estructuras de Datos Avanzadas/Las Pistas de Azul/cases/caso5.1.out
