+++ Estructuras de Datos Avanzadas/Las Pistas de Azul/cases/caso5.2.in
@@ -1,4 +1,5 @@
-15 12 3
+15
+12 3
 3 2 13 7 9 9 7 6 11 12 3 9 2 12 7 5 11 12 7 0 8 6 7 3 1 10 13 0 0 12
 5 13 5 14 7 9 6 11 11 1 1 14 1 14 4 14 8 8 7 11 13 6 13 13 4 13 7 13 10 5
 2 7 1 12 12 11 1 11 14 1 2 3 5 0 1 13 8 13 10 0 7 0 4 3 2 3 6 10 3 8
