+++ Estructuras de Datos Avanzadas/Las Pistas de Azul/cases/caso5.2.out
