+++ Estructuras de Datos Avanzadas/Las Pistas de Azul/examples/sample.in
@@ -1,0 +1,1 @@
+../cases/caso1.1.in
