+++ Estructuras de Datos Avanzadas/Las Pistas de Azul/settings.json
@@ -1,0 +1,28 @@
+{
+  "title": "Las Pistas de Azul",
+  "source": "Curso OMI (Orac)",
+
+  "limits": {
+    "TimeLimit": 1000,
+    "MemoryLimit": 134217728,
+    "InputLimit": 67108864,
+    "OutputLimit": 67108864,
+    "ExtraWallTime": 0,
+    "OverallWallTimeLimit": 60000
+  },
+
+  "validator": {
+    "name": "token-caseless",
+    "limits": {
+      "TimeLimit": 1000
+    }
+  },
+
+  "misc": {
+    "alias": "COMI-Las-Pistas-de-Azul",
+    "visibility": "public",
+    "languages": "all",
+    "email_clarifications": 0,
+    "admin-groups": ["workshop-2017"]
+  }
+}
