+++ Estructuras de Datos Avanzadas/Las Pistas de Azul/testplan
@@ -1,0 +1,10 @@
+caso1.1 20
+caso1.2 0
+caso2.1 20
+caso2.2 0
+caso3.1 20
+caso3.2 0
+caso4.1 20
+caso4.2 0
+caso5.1 20
+caso5.2 0
