+++ Estructuras de Datos Avanzadas/Las Pistas de Azul/tests/tests.json
@@ -1,0 +1,12 @@
+{
+	"solutions": [
+		{
+			"filename": "../solutions/solution.cpp",
+			"verdict": "AC"
+		}
+	],
+	"inputs": {
+		"filename": "test-validator.py"
+	},
+	"max_score": 100
+}
