+++ problems.json
@@ -1,7 +1,7 @@
 {
   "problems": [
     {
-      "path": "Programacion Basica/Palindromos"
+      "path": "Estructuras de Datos Avanzadas/Las Pistas de Azul"
     }
   ]
 }
